requirements.txt
openai
sentence-transformers
chromadb
python-dotenv
numpy
requests